Create models module.
